+++ artillery-core/Cargo.toml
@@ -19,6 +19,7 @@ chrono = { version = "0.4", features = ["serde"] }
 rand = "0.7.3"
 mio = { version = "0.7.0-alpha.1", features = ["os-poll", "udp"] }
 futures = "0.3"
+pin-utils = "0.1.0-alpha.4"
 libp2p = "0.16.0"
 bastion-executor = "0.3.4"
 lightproc = "0.3.4"
@@ -31,16 +32,10 @@ clap = "2.33.0"
 pretty_env_logger = "0.4.0"
 once_cell = "1.3.1"
 criterion = "0.3.1"
-trybuild = "1.0"
+kaos = "0.1"
 bastion = "0.3"
 
 
 [[test]]
-name = "epidemic-state-change-tail-follow"
-path = "tests/failpoints/epidemic-state-change-tail-follow.rs"
-required-features = ["fail/failpoints"]
-
-[[test]]
-name = "epidemic-periodic-index"
-path = "tests/failpoints/epidemic-periodic-index.rs"
-required-features = ["fail/failpoints"]
+name = "kaos"
+path = "kaos-tests/launcher.rs"
