+++ artillery-core/examples/cball_ap_cluster.rs
@@ -48,8 +48,7 @@ fn main() {
     };
 
     // Configure our cluster node
-    let (cluster, _cluster_listener) = ArtilleryAPCluster::new(ap_cluster_config).unwrap();
-    let ap_cluster = Arc::new(cluster);
+    let ap_cluster = Arc::new(ArtilleryAPCluster::new(ap_cluster_config).unwrap());
 
     // Launch the cluster node
     run(
@@ -60,7 +59,7 @@ fn main() {
             let ap_events = ap_cluster.clone();
 
             // Detach cluster launch
-            let cluster_handle = spawn(async move { ap_cluster.launch().await }, cluster_stack);
+            let cluster_handle = spawn_blocking(async move { ap_cluster.launch().await }, cluster_stack);
 
             // Detach event consumption
             let events_handle = spawn_blocking(
