+++ artillery-core/kaos-tests/epidemic-state-change-tail-follow.rs
@@ -1,0 +1,14 @@
+extern crate pretty_env_logger;
+
+#[macro_use]
+extern crate log;
+
+#[macro_use]
+mod chaos;
+
+use chaos::*;
+
+fn main() {
+    cluster_init!();
+    chaos_unleash!("epidemic-state-change-tail-follow-fp");
+}
