+++ artillery-core/kaos-tests/mod.rs
@@ -1,0 +1,6 @@
+extern crate pretty_env_logger;
+
+#[macro_use]
+extern crate log;
+
+mod chaos;
