+++ artillery-core/src/cluster/ap.rs
@@ -4,8 +4,11 @@ use crate::service_discovery::mdns::prelude::*;
 
 use lightproc::prelude::*;
 
-use std::sync::Arc;
+use std::{cell::Cell, sync::Arc};
 use uuid::Uuid;
+use futures::{FutureExt, select};
+use pin_utils::pin_mut;
+
 
 #[derive(Default, Clone)]
 pub struct ArtilleryAPClusterConfig {
@@ -19,6 +22,7 @@ pub struct ArtilleryAPCluster {
     config: ArtilleryAPClusterConfig,
     cluster: Arc<Cluster>,
     sd: Arc<MDNSServiceDiscovery>,
+    cluster_ev_loop_handle: Cell<RecoverableHandle<()>>,
 }
 
 unsafe impl Send for ArtilleryAPCluster {}
@@ -27,20 +31,20 @@ unsafe impl Sync for ArtilleryAPCluster {}
 pub type DiscoveryLaunch = RecoverableHandle<()>;
 
 impl ArtilleryAPCluster {
-    pub fn new(config: ArtilleryAPClusterConfig) -> Result<(Self, RecoverableHandle<()>)> {
+    pub fn new(config: ArtilleryAPClusterConfig) -> Result<Self> {
         let sd = MDNSServiceDiscovery::new_service_discovery(config.sd_config.clone())?;
 
         let (cluster, cluster_listener) =
             Cluster::new_cluster(config.node_id, config.cluster_config.clone())?;
 
-        Ok((
+        Ok(
             Self {
                 config,
                 cluster: Arc::new(cluster),
                 sd: Arc::new(sd),
-            },
-            cluster_listener,
-        ))
+                cluster_ev_loop_handle: Cell::new(cluster_listener)
+            }
+        )
     }
 
     pub fn cluster(&self) -> Arc<Cluster> {
@@ -56,6 +60,23 @@ impl ArtilleryAPCluster {
     }
 
     pub async fn launch(&self) {
+        let (_, eh) = LightProc::recoverable(async {}, |_|(), ProcStack::default());
+        let ev_loop_handle = self.cluster_ev_loop_handle.replace(eh);
+
+        // do fusing
+        let ev_loop_handle = ev_loop_handle.fuse();
+        let discover_nodes_handle = self.discover_nodes().fuse();
+
+        pin_mut!(ev_loop_handle);
+        pin_mut!(discover_nodes_handle);
+
+        select! {
+            ev_loop_res = ev_loop_handle => { dbg!(ev_loop_res); ev_loop_res.unwrap() },
+            _ = discover_nodes_handle => panic!("Node discovery unexpectedly shutdown.")
+        };
+    }
+
+    async fn discover_nodes(&self) {
         self.service_discovery()
             .events()
             .iter()
